requirements.txt
jinja2
markupsafe
wtforms
werkzeug
dominate
flask
flask-bootstrap
flask-wtf
itsdangerous
visitor
flask-nav
pyserial
